+++ demo/symfony8/config/bundles.php
@@ -3,12 +3,15 @@
 declare(strict_types=1);
 
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
+use Nowo\FormKitBundle\NowoFormKitBundle;
 use Nowo\TwigInspectorBundle\NowoTwigInspectorBundle;
+use Nowo\UiKitBundle\NowoUiKitBundle;
 use Nowo\WorkflowBundle\NowoWorkflowBundle;
 use Symfony\Bundle\DebugBundle\DebugBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Bundle\WebProfilerBundle\WebProfilerBundle;
+use Twig\Extra\TwigExtraBundle\TwigExtraBundle;
 
 return [
     FrameworkBundle::class         => ['all' => true],
@@ -18,4 +21,7 @@ return [
     DoctrineBundle::class          => ['all' => true],
     NowoWorkflowBundle::class      => ['all' => true],
     NowoTwigInspectorBundle::class => ['dev' => true, 'test' => true],
+    NowoUiKitBundle::class         => ['all' => true],
+    NowoFormKitBundle::class       => ['all' => true],
+    TwigExtraBundle::class         => ['all' => true],
 ];